feat: new version please new60

+++ publish.sh
@@ -9,19 +9,19 @@ echo "ver: $VERSION $TAG"
 ls && ls me-so-solly
 
 if [ "$?" != "0" ]; then
-  echo "failed to deploy, removing the already created tag";
+  echo "failed to deploy";
 
   mkdir ~/.ssh/ && echo -e "Host github.com\n\tStrictHostKeyChecking no\n" > ~/.ssh/config
 
-  echo fetch
   git fetch --all
+  GITHUB_TAG=`git --no-pager tag | grep $TAG`
 
-  echo tag
-  git --no-pager tag
+  echo $GITHUB_TAG
 
-  git --no-pager tag | grep $TAG
-
-  git tag -d $TAG && \
-  git push --delete origin $TAG
-  exit 1;
+  if [ "$GITHUB_TAG" != "" ]; then
+    echo "Removing the already created tag: $TAG";
+    git tag -d $TAG && \
+    git push --delete origin $TAG
+    exit 1;
+  fi
 fi
